update : Modular Navbar
